Change at-columns to bigInteger()

+++ persistence/migrations/0030_accountAdministrators.ts
@@ -8,8 +8,8 @@ export const up = async (knex: Knex): Promise<void> =>
     table.integer('userId')
 
     table.boolean('deleted').nullable()
-    table.integer('createdAt')
-    table.integer('updatedAt').nullable()
+    table.bigInteger('createdAt')
+    table.bigInteger('updatedAt').nullable()
 
     table.foreign('accountId')
       .references('id')
