+++ persistence/migrations/0050_polls.ts
@@ -9,8 +9,8 @@ export const up = async (knex: Knex): Promise<void> =>
     table.string('name')
 
     table.boolean('deleted')
-    table.integer('createdAt')
-    table.integer('updatedAt')
+    table.bigInteger('createdAt')
+    table.bigInteger('updatedAt')
 
     table.foreign('accountId')
       .references('id')
