+++ persistence/migrations/0110_messageMetas.ts
@@ -17,8 +17,8 @@ export const up = async (knex: Knex): Promise<void> =>
     table.string('type') // TODO: Заменить на enum
 
     table.boolean('deleted').nullable()
-    table.integer('createdAt')
-    table.integer('updatedAt').nullable()
+    table.bigInteger('createdAt')
+    table.bigInteger('updatedAt').nullable()
 
     table.foreign('accountId')
       .references('id')
