+++ pincodes-service/migrations/0030_pinCodes.ts
@@ -11,8 +11,8 @@ export const up = async (knex: Knex): Promise<void> =>
     table.json('payload')
 
     table.boolean('deleted').nullable()
-    table.integer('createdAt')
-    table.integer('updatedAt').nullable()
+    table.bigInteger('createdAt')
+    table.bigInteger('updatedAt').nullable()
 
     table.foreign('tenantId')
       .references('id')
